update config to allow for multiple rows
