+++ packages/esm-ward-app/src/ward-patient-card/ward-patient-card.tsx
@@ -1,12 +1,18 @@
 import React from 'react';
+import { useParams } from 'react-router-dom';
 import { type WardPatientCardProps } from '../types';
-import WardPatientCardRow from './ward-patient-card-row.component';
+import { usePatientCardRows } from './ward-patient-card-row.resources';
 import styles from './ward-patient-card.scss';
 
 const WardPatientCard: React.FC<WardPatientCardProps> = (props) => {
+  const { locationUuid } = useParams();
+  const patientCardRows = usePatientCardRows(locationUuid);
+
   return (
     <div className={styles.wardPatientCard}>
-      <WardPatientCardRow {...props} className={styles.wardPatientCardHeader} />
+      {patientCardRows.map((WardPatientCardRow) => (
+        <WardPatientCardRow {...props} />
+      ))}
     </div>
   );
 };
